bug: fixed rare bug for nnzs == nspin
When systems were parallelizes such that a processor
only had nnzs == nspin, then the sparse IO routines where
doing it wrong since it would swap the dimensions.

I have now added a field in the class_SpData2D to clarify
which dimension is the sparsity dimension.
I am also guessing it could be vital to build H/DM matrices
for SOC+NC such that the sparsity dimension is transposed
for faster access.

Thanks to Javier Junquera for finding this.