maint: changed sign and diag-spiral code
Moved from pure real arrays to complex arrays.
This is much easier to read and clarifies some things.

This uncovered that a sign-difference between the original
spiral code and the current Siesta base.
If one compares the diagk code and the original spin-spiral
code there was a mismatch between the phases. This should
now be fixed and meant reversing signs in vmatsp and rhoofdsp
as well.

The code is still very much un-tested since we don't have
a decent test for this.