Better processing of the valence charge for ps generation
The header information in psf files records information about the
valence configuration for pseudopotential generation. Some generators
(such as ATOM) cannot pseudize multiple shells for the same angular
momentum.  When semicore states are present, ATOM uses an ionic
configuration, with a reduced valence charge. The legacy header,
having one field per 'l', is still appropriate. Other programs (such
as ONCVPSP) can deal with multiple shells per 'l' in the valence, but
the psf header cannot contain the full information, and the total
valence charge at the time of pseudopotential generation has to be
handled in a special way, beyond adding up the contributions of the
different sections of the header.

An extra field in psf files that records the total valence charge for
pseudopotential generation was present already ('gen_zval', for the
needs of the VCA tools), but, due to a bug, Siesta was not processing
properly this field, even if set.

The 'gen_zval' value can be significant. Its difference from the
effective valence of the pseudo-ion is used by default as an 'extra
charge' for PAO generation, and it also determines the charge used for
KB projector generation.

The above was an issue when using psf files generated from PSML files
with semicore states, but with a non-ionic configuration. In this
case, Siesta set incorrectly the 'gen_zval' value, forcing the use of
a non-zero effective charge for PAO generation. This led typically to
shorter orbitals.

Siesta now processes correctly the 'gen_zval' field (if present). Note
that in this version the user still has to provide a PAO.Basis block
with the correct suite of shells if semicore states are included in
the valence. And, of course, the psf file should correctly reflect the
proper valence charge, if appropriate.