Changed TS dq tolerance, now per SCF step.
A report was made about the SCF not converging.
Clearly the user did not read the message since it clearly
stated that there was abnormal charge.

I have now made the checks be in the SCF loops which is probably
easier to understand. This may end up in situations where
the SCF never stops due to dQ being too high.
There is no good way around this, I think.
Perhaps this scheme requires another iteration.