enh: enabled mesh distribution based on lattice vector sizes
The old grid-distributions only allowed factors of
2, 3 and 5. This was limiting in terms of the divisions
for the mesh.

This commit enables one to divide the mesh according to
the lengths of the vectors that are to be divided.

I.e. if the 3rd lattice vector is *much* longer than
the 2nd then we should have more divisions on the
3rd compared to the 2nd. This was never taken into
account and could also limit the scalability when
exa-scale processors are used.
Now there is no preferred factor, it will try
and divide the processors such that:

   Nz/Ny ~ |3| / |2|

to make it as even as possible.