Add python script for GitLab releases.
The main purpose of the script is to create a SIESTA release on GitLab.

The script may also be used to delete a release from GitLab or to replace the
description (text) of an existing release.  These two additional features will
work for any GitLab release (not only for SIESTA releases).